chore: refactor and review
